Add colliders and player character
Added box colliders to floor and wall.
Added player character from "Ultimate Modular Women Pack" from Quaternius. Copied over raycast source gameobjects to the witch (player character). Added mesh colliders, movement controller script, interaction manager, script, and rigidbody to the witch. There is currently a problem with the witch character hovering when the "W" key is pressed and UpperStepSource is on 0.521 Y position or higher. Further description of the issue is on Discord.