Add red arrow above interactable when raycast hits
I had to create a new EndDayTable parent gameobject and I copied and pasted over the components of EndDayTable and moved the child gameobjects.

This was because or some reason the EndDayTable was completely off in its position. For example, if the transform said it was in the x position of 20, then the actual object in the x position of 20 was a distance to the left and the object in the same x position as the EndDayTable was actually in the x position of 30.

I believe I could see the discrepancy in its position once I changed from "Center" to "Pivot" in the top left (after the old EndDayTable was empty of child objects but still had its components and I only added an empty cube on there).