Fix Inventory UI text squash and bigger dropdown
To fix the Inventory UI squash I changed the y from 2 to 2.625 of all the child gameobjects (except the Markup gameobject and the inactive Static_Markup gameobject) of InventoryItemPrefab. My guess of the cause of the squished text is caused by the scaling of the parent gameobjects. For example the Content gameobject under the Viewport which is under the Inventory Scroll View which is then under the Inventory Panel has a 10.5, 4, 1 scaling (x, y, z). The Inventory Scroll View and Inventory Panel gameobjects have their own scaling too.

What I changed on InventoryItemPrefab to get bigger dropdown box: 50 markup height, 56 height content, 40 height item. I also doubled the font size in Label and Item Label.