Add tax text and add refs to database (again)
For some reason, some of my work in the Unity scene (even though I committed) keeps getting rolled back. (Same thing happened at least once before.)

In this instance it was the references in the GameManager script missing tax text, daily sales number, and money count. Also I had to readd the TaxTitle and TaxAmount gameobjects containing the text component.

When I switched to main branch and switched back to GongBingDev, there was even more of a rollback with the "Daily Sales: Number" text missing as well. (I don't know if the "Add 100 Daily Sales" button was missing too but it might have been as well). Closing Unity and reopening it again brought that back but the stuff in the prior paragraph was still missing so hence this commit.