Change store and edit money value in title screen
I deleted a lot of assets in one of my commits and these assets were used in the title scene (which caused errors) so I copied the store from the store scene to use in the title screen. I couldn't copy the interior since since the pedestals were connected to scripts.

I also changed the MoneyValue gameobject's Text component's Horizontal Overflow from Wrap to Overflow to prevent the money being split into two rows (with a "0" being shown on the 2nd line with 1000.00 money).