Fix the "Money Made" being 0 in EndDayScreen
The "money made being 0" issue was because of "dailySales = 0;" being called in IncrDay() function which is before the EndDay() function that displays the money made. So I moved "dailySales = 0" to the GameManager script after the EndDay() function.

After that the money made never reset (kept going up) so I made changes to the timer active, but then realized it was meant to be a one time trigger, so I undid the changes. And then it randomly fixed itself? (Money reset again.) I have 0 idea what happened.