Fix wage calculation in net profit
GameManager.cs: I added a "-" in front of wagesPaidAmount since the variable is negative. Before my change, if you subtracted wagesPaidAmount (which is a negative number), you would have the wages be added to the moneyMade (since a negative and a negative makes a positive).

JsonDatabaseOperations.cs: I changed "currentPlayer.active[i] = 1;" from 0 because its easier for me to test with all the items already on the tables.