Add sound effects
Added:
1) Sound when items are purchased from cauldron (differences in 1 and 10 items and successful and unsuccessful buys)
2) Sounds for entering different game states
3) Button clicks
4) Items being placed + sold

For some reason, the buttons of ExitButton (1) and SettingsButton had to have SoundManager.ButtonClickSound added separately. This means I couldn't include those two buttons with my selection of all the other buttons when I tried to add the sound to the On Click () or else the settings menu and/or the exit button of the settings would not work. Same with the TitleScreenButton, it replaces another onClick function with a duplicate of the other OnClick if there's 2 existing OnClicks functions (along with the adding on the ButtonClick sound).

I ended up only batch selecting buttons I knew had the same amount of OnClick functions just to be safe and applying the OnClick() button click sound to the rest manually one by one. (I think the OnClick one or none functions could be batch selected without issues but I didn't want to risk it.)