Set Loans Panel2 gameobject active
For some reason the LoansMenu gameobject has to be active in the Store scene in Edit mode to have it open in Play mode in one "F" press (after the tutorial UI is closed). Two "F" presses are required (after the tutorial UI is closed) if its inactive in the Store scene in Edit mode.