Fix LoansMenu "F" press bug
When setting the LoansMenu gameobject inactive in the Hierarchy, the LoansMenu does open once you press F, however it immediately closes right after because of the SetActive(false) in the Start() function. Which is why the "F" button had to be pressed twice to work (excluding the "F" press to open the tutorial pop-up UI).