Add customer picking up item animation
Also, the Animator on the Customer prefab is now just one Walking state. I just used the witch walking animation which loops.

I created a new layer with an AvatarMask (with legs excluded) and adjusted the weight of the layer to 0.75 so that the walking animation still played when the customer is doing item pickup animations and still moving.