Add day menu changes
The day menu can be pulled up at any time to show the previous day's statistics.

In addition, it has a "Start Selling" button and an "End Day" button.

This means that starting the selling phase and ending the day doesn't happen instantly when you interact with the End Day table.