Fix customer purchasing animation bug
Had to fix the same bug that I believe the "Make non-buying customer not play animation" commit had (with my belief based on the commit title) where the customer was stretching their hand out while item wasn't being sold.

I did this with a rework of the animation system by playing the customer interact animation within the function that calls the SellItem() function . I referenced the customer through the box collider in front of the table.

Within the Store scene, each table now has a TableTrigger gameobject with a box collider in front of the table and the Customer Interact Handler script attached with each script having its own unique table merch ID.